refactor: modulize json processing module
