+++ python/jskiner/folder/__init__.py
@@ -1,0 +1,3 @@
+from .main import JsonFileProcessor
+
+__all__ = ["JsonFileProcessor"]
