+++ python/jskiner/folder/filter.py
